unwrap array which is always of length 0 or 1
parseFormat's second argument was always an empty array, since its sole callsite, in parseFragment, checked that the array in question was empty before invoking it.

parseFormat invoked parseFragment with an array of length 1 as its second argument (we know it was of length 1 because it resulted from calling fmtStack.concat(string), where fmtStack was empty per the previous paragraph).

No other caller of parseFragment passed the second argument, so it would always default to the empty array in that case.

So there's no reason for the argument to be an array at all, rather than just a string or undefined.